[feat,chore,refactor,test] Add amova ci and negative variance adjustment, v0.5
[feat]

add amova confidence interval using bootstrap percentile method, normal
approximation, bootstrap basic method (reverse percentile method). When block
bootstrapping is enabled, the results from all 3 of these methods will be
reported by default.

add missing data check dmat->has_drop for neighbor joining

add negative variance adjustment and alternative results via setting negative
variance components to 0 and recalculating the phi statistics.

[chore]

change dxy and amova result printing for double values from %f (float) to %.17g (double)

deprec mathUtils

add cross-platform __FUNCTION__ identifier

change the name of --drop-pairs to --allow-missing-pairs for clarity

[refactor]

improve block bootstrapping samples printing

rename nsites variables in em.cpp for clarity (does not affect the results)

[test]

add block bootstrapping test case test18

Update test results with including negative variance adjusted results

+++ paramStruct.cpp
@@ -181,7 +181,7 @@ void estimate_memory_needed(paramStruct* pars, vcfData* vcfd) {
                 ++ncontigs;
                 ++contigidx;
             } else {
-                if (contigidx != ncontigs - 1) {
+                if ((int) contigidx != ncontigs - 1) {
                     ERROR("Contig names in the alleles file %s are not sorted. Please make sure they are sorted by contig name.", fn);
                 }
             }
@@ -254,7 +254,7 @@ void estimate_memory_needed(paramStruct* pars, vcfData* vcfd) {
     alleles->d[currblock] = tmp;
 
     // we can also calculate the expected j as:
-    DEVASSERT(j == (((int)nrow % 16) * 4));
+    DEVASSERT((int) j == (((int)nrow % 16) * 4));
     DEVASSERT(i == nrow);
     DEVASSERT(alleles->cposidx->len == alleles->cnames->len);
     DEVASSERT(alleles->pos->len == nrow);
@@ -485,6 +485,8 @@ paramStruct* paramStruct_init(argStruct* args) {
     pars->alleles_contigidx = -1;
     pars->a1a2[0] = -1;
     pars->a1a2[1] = -1;
+    // pars->b1b2[0] = -1;
+    // pars->b1b2[1] = -1;
     pars->formulaTokens = NULL;
     pars->ibd = NULL;
     pars->DATETIME = NULL;
